refactor(#383): split HttpServer into DiagnosticsEndpoints + OperatorEndpoints + AdminEndpoints
Closes #383.

HttpServer.cs was a 1204-LOC 'everything endpoint' with 19 routes and a
13-parameter constructor that mixed three operational concerns. Split
into three handler classes per the issue, each constructed with only
its own dependencies:

  - DiagnosticsEndpoints (read-only):
      /health/live, /health/ready, /metrics, /sessions, /sessions/{id},
      /firms
  - OperatorEndpoints (channel/instrument-scoped mutations):
      /channel/{ch}/snapshot-now, bump-version, trade-bust;
      /admin/post-trade/bust;
      /admin/instruments/{securityId}/{phase,halt,resume}
  - AdminEndpoints (host-wide admin):
      /admin/daily-reset, /admin/post-trade/eod-export,
      /admin/channels/{ch}/snapshot{,/force,/reset,/validate}

Shared parsing/JSON helpers (TryParseLong/ULong, JsonEscape,
AdminJsonOptions, PhaseChangeTimeout, LocalMktDateEpoch) live in
HttpInternals.cs.

HttpServer now owns only routing + Kestrel bootstrap (1204 -> 135 LOC).
Its public constructor signature is unchanged so ExchangeHost and
existing tests work without modification. All 1,311 tests pass; dotnet
format clean.
